Changes related to Windows packaging
When using PyInstaller, the 'skytemple' distribution is not installed,
so we can't use it to get the version or the entrypoints. As such
we have a special solution for loading modules on Windows for now.

Will probably extend that later to also lookup files in a special
'modules' directory of the install dir under Windows, so Windows users
can also install third-party modules.